changes made to rts game by Nicholas Alfonso
So the changes I made in the unity game was first giving the player and enemy units different colors to differentiate each other with player units being blue and enemy units being red. To make the game somewhat easier and  functional was editing the player units so that they are a bit stronger during the start but can still be defeated by enemy units and to help the player get more units they will start off with 100 gold to get as many as they need to wipeout 7 enemy units to beat the game. I even fixed how after an enemy is defeated you can have your units grab the weapon pick up the enemy dropped  and even fixed the healing. But quite honestly it wasn't really easy since my computer is a bit on the slow side but I tried to get things done so it at least looks somewhat decent but hopefully all goes well.