My skeleton fighter Bonestrike by Nicholas Alfonso
So for my fighter, I went with a skeleton sprite that I found in itch.io and gave him the name Bonestrike who fights his foes with a magic flail. While making Bonestrike I tried to opt for a more balanced character something akin to most fighting game mascot characters like Ryu from street fighter and make it so that the player can easily get a good grip as to whom they are playing as and can easily master the character to get a good sense of an introduction to fighting games. When it comes to the controls for Bonestrike I made it so that you can use the basic left and right arrow keys for horizontal movement and the player could dash using the right side shift key for short time and for the attacks I used the a s d keys for light, medium, and heavy attacks for Bonestrike. To make Bonestrike  be a balanced fighter I made it pretty simple for his attacks to do the appropriate amount of damage for which attack button is pressed and I made him so that is equal in speed and is perfectly mid-weight enough for handling the little boney boy. My only challenge while creating my fighter was dealing with the animation in where the sprites I had didn't have much to make some animation that compliments to an action. The best example I could give is where sometimes certain animations like the dashing and running are a bit hard to tell apart aside from speed as well as some animations looking a bit weird but there is always room for improvement even if animating sprites is a bit difficult to do at the first try but I believe I can get better later and just quickly want to mention this and yes I could've used AI to help me make a sprite sheet with more better  sprites but I felt unsure since it would be a bit tricky to slice the sprite sheet in unity. Overall, I tried to do the best I could right now to make it look ok and I hope that I works without any sort of issues but only time will tell and if later there is an opportunity to fix it then I would definitely take the time to fix any errors but until then this is the best I have and if I did everything correctly hopefully it works and you get this without any issues. Well that is pretty much it for now so I will see you next time.