Unity Brawler Edits and Changes by Nicholas Alfonso
So for todays assignment, I was pretty much tasked to fix and edit any thing about my 2D fighting game character and add a custom health bar for my character and make a script to go along with it. For those curious my fighter is a skeleton warrior named Bonestrike who fights enemies with a magic mace  and to keep things brief he is designed to be more of a balanced fighter so that it could be easy for newcomers to the fighting-game genre but I felt like I could change him up a bit. For starters, I decided to fix his hit box since before I felt was a bit short in terms of range so I made it more bigger to compensate but still having it be the perfect amount of range to still land attacks. Another change I made was fixing his gravity and having him not hover and keep falling because before he used to be suspended in mid-air once you start the game and after he attacks he kept doing his falling animation because the sprite couldn't find the ground but that was easily fixed by resetting the capsule collider and moving the ground check a bit and everything was good afterwards. Now, for the major change being my custom health bar and for that I created a UI and placed it inside my fighter and used a sprite found in our UI sprite folder from github and made it an image so that it would be easy since its a simple square bar that goes down as you take damage and I made it show up on the upper left corner of the screen just like how it is with most 2D fighting games. The only obstacle was creating the script of course but with some slight help from the professor and some minimal AI use to check and make sure I'm doing things correctly I was able to make a health bar script that should be functioning and added it onto my fighter. So that is all that there is as to what changes where made and again the only thing I hope for is that everything is working fine but it should work but let me know what you think and until then see you next time.